requirements.readthedocs.txt
sphinx<=1.6.7
numpydoc>=0.5
sphinx_rtd_theme>=0.1.7
scipy>=0.8
algopy>=0.4
statsmodels>=0.6
hypothesis>=3.6
pytest
